Unicomp New Model M Pause/Break key fix
The PS/2 controller code on Unicomp's New Model M has a logic design
flaw whereby it outputs E0 77 (effectively substituting a press of Num
Lock) instead of E0 7E E0 F0 7E (substituting a press and release of
Scroll Lock, which is correct for Break, given its legacy from the IBM
5150). Added this into the function that maps E0 scancodes to correct
for Unicomp's mistake. Tested, and does not interfere with actually
pressing Ctrl + Num Lock, which executes a Pause, as it should.