NEW FEATURE: ibmpc_usb converter lock state LEDs
Restructured lock state LED routine to update both the keyboard and any
indicators wired to the converter itself.

This deliberately does not use the Quantum indicator LED tools because
that interfered with the more TMK-relevant routine for translating the
USB HID LED update byte to IBM format and sending it to the keyboard,
and also because TMK does not have the Quantum tools, and I would
rather that the TMK, QMK, and Vial versions of this converter code were
reasonably similar to make it more straightforward to keep all three
updated somewhat in tandem. Consequently, pin assignments for LEDs
wired to the converter are in led_pins.h for the time being.