Refactored ibmpc_usb LED setter to use Quantum
On second thoughts, it feels dumb to write a controller pin handler
when one already exists in QMK as it is, so I've gone back to the
original strategy and used the built-in led.h writePin instead. It's
not a huge amount of differentiation between TMK and QMK versions of
the converter anyway, so I doubt it will complicate matters much (it
just means some of the TMK code won't be needed in the QMK/Vial version
and the writePin chunk that's unique to QMK/Vial needs to not be
interfered with when porting across updated TMK code in future).