Refactored in response to feedback re. separation
My intention with splitting up ibmpc_host_set_led() into two functions
was to ensure that the bit order of the USB HID LED update byte was not
reconfigured to IBM order unless it was actually going to be used.
However, it has been suggested that splitting the check for whether the
keyboard can receive the LED byte and the subsequent sending of the LED
byte into separate functions in some way implies that they are less
contextually connected and inter-reliant, which could hypothetically
lead to a misunderstanding and potentially the addition of code that
does unwise things with these newly separate functions (e.g. you can't
send the LED byte without sending the check byte immediately prior, you
can't send multiple LED bytes after only ascertaining once that the
keyboard is capable of receiving them, and so forth).

To reduce the likelihood of this interpretation, I've removed the case-
specific functions for sending the two bytes from ibmpc and called
ibmpc_host_send() directly from the converter side to make it clearer
that the LED byte should only be sent immediately after the check byte.